requirements.txt
aioice==0.9.0
aiortc==1.9.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.7.0
asttokens==3.0.0
attrs==24.2.0
av==12.3.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
comm==0.2.2
cryptography==44.0.0
debugpy==1.8.10
decorator==5.1.1
distro==1.9.0
dnspython==2.7.0
executing==2.1.0
gitdb==4.0.11
GitPython==3.1.43
google-crc32c==1.6.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
ifaddr==0.2.0
imageio==2.36.1
imageio-ffmpeg==0.5.1
ipykernel==6.29.5
ipython==8.30.0
jedi==0.19.2
Jinja2==3.1.4
jiter==0.8.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
moviepy==2.1.1
narwhals==1.17.0
nest-asyncio==1.6.0
numpy==1.26.4
openai==1.57.2
packaging==24.2
pandas==2.2.3
parso==0.8.4
pillow==10.4.0
platformdirs==4.3.6
proglog==0.1.10
prompt_toolkit==3.0.48
protobuf==4.25.5
psutil==6.1.0
pure_eval==0.2.3
pyarrow==18.1.0
pycparser==2.22
pydantic==2.10.3
pydantic_core==2.27.1
pydeck==0.9.1
pydub==0.25.1
pyee==12.1.1
Pygments==2.18.0
pylibsrtp==0.10.0
pyOpenSSL==24.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pywin32==308
pyzmq==26.2.0
referencing==0.35.1
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
setuptools==75.6.0
six==1.17.0
smmap==5.0.1
sniffio==1.3.1
stack-data==0.6.3
streamlit==1.33.0
streamlit-webrtc==0.47.9
tenacity==8.5.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
watchdog==6.0.0
wcwidth==0.2.13
